Animation Refactoring
- renamed GG animation types to Ruby
- added Ruby_AnimationSetup to Ruby_Animation
- added vdp location to Ruby_Animation
- removed all the batched and meta sprite animations
- udpated Load/Setup/Update animation functions

Misc
- added lots of logging
- disabled the RightScroll command runner as there's something wrong with it

+++ project/source/engine/vdptile_manager.c
@@ -13,9 +13,9 @@ void VDPTileManager_Init(void)
 	VDPTileManager_backgroundTileLocation = BACKGROUND_TILES_START_INDEX;
 }
 
-u8 VDPTileManager_LoadSpriteTiles(const u8* tileData, 
-								  u8 tileCount,
-								  u8* vdpLocation)
+u8 VDPTileManager_LoadSpriteTiles(const u32* tileData, 
+								  u16 tileCount,
+								  u16* vdpLocation)
 {
 	u16 currentSpriteTileLocaton = VDPTileManager_spriteTileLocation;
 
@@ -35,8 +35,8 @@ u8 VDPTileManager_LoadSpriteTiles(const u8* tileData,
 	return *vdpLocation;
 }
 
-u8 VDPTileManager_ReserveSpriteTilesArea(u8 tileCount, 
-										 u8* vdpLocation)
+u8 VDPTileManager_ReserveSpriteTilesArea(u16 tileCount, 
+										 u16* vdpLocation)
 {
 	u16 currentSpriteTileLocaton = VDPTileManager_spriteTileLocation;
 
@@ -53,7 +53,7 @@ u8 VDPTileManager_ReserveSpriteTilesArea(u8 tileCount,
 
 }
 
-u16 VDPTileManager_LoadBackgroundTileset(const u8* tileData, 
+u16 VDPTileManager_LoadBackgroundTileset(const u32* tileData, 
 										 u16 tileCount,
 										 u16* vdpLocation)
 {
