+++ project/source/engine/vdptile_manager.h
@@ -7,14 +7,14 @@
 
 void VDPTileManager_Init(void);
 
-u8 VDPTileManager_LoadSpriteTiles(const u8* tileData, 
-								  u8 tileCount,
-								  u8* vdpLocation); 
+u8 VDPTileManager_LoadSpriteTiles(const u32* tileData, 
+								  u16 tileCount,
+								  u16* vdpLocation); 
 
-u8 VDPTileManager_ReserveSpriteTilesArea(u8 tileCount, 
-										 u8* vdpLocation);
+u8 VDPTileManager_ReserveSpriteTilesArea(u16 tileCount, 
+										 u16* vdpLocation);
 
-u16 VDPTileManager_LoadBackgroundTileset(const u8* tileData, 
+u16 VDPTileManager_LoadBackgroundTileset(const u32* tileData, 
 										 u16 tileCount,
 										 u16* vdpLocation);
